requirements.txt
llama-index-core>=0.12.41
llama-index-readers-file>=0.4.0
python-dotenv>=1.0.0
openai>=1.0.0